update trx01 qaqc flags for automated qaqc
Currently, QAQC flags are set as -111 and -112 if CO2 or CH4 are outside a set range, respectively. These flags are not mentioned in the pipeline README. I propose that they should be changed to -1 (data manually removed), which is generally used by the bad_data_fix.r script, but I think would fit for this situation.